+++ NEWS.md
@@ -1,6 +1,7 @@
 ## DTSg v0.1.3.9000
 
 * Added `swallow` argument to the `new` method: It allows for a more ressource efficient object creation
+* Added `drop` argument to the `values` method: It allows for a ressource efficient destruction of a `DTSg` object while preserving its *values*
 * Added `timestamps` field providing the total number of timestamps
 * Added means to measure code coverage with the help of `covr`
 * Removed `xts` from suggested packages list (already comes along with `dygraphs`)